test: move jax mocks and utils to test folder
